Zufallsbereich ausgegliedert
- Datei zufallsbereich hinzugefügt.
- Funktion aus exel-constants herausgenommen
- Fragment zufallsbereich in excel.ts hinzugefügt
